requirements.txt
open3d
plyfile
opencv-python
trimesh==2.35.39
networkx==2.6.3
scipy==1.7.3
tensorboardX
easydict
tqdm
h5py
matplotlib
elevate
numpy-quaternion
imageio
